rpi_control renamed
